requirements.txt
numpy
tqdm
pillow
opencv
torch
torchvision
tensorboard
h5py